requirements.txt
streamlit
tensorflow
numpy
opencv-python-headless
pillow
requests
streamlit-lottie
openai
matplotlib